requirements.txt
PyGithub>=2.1.1
tqdm>=4.62.3
